Move sendForm components to sendFlow
