+++ src/components/SendFlow/MultisigAccount/FormPage.test.tsx
@@ -11,7 +11,7 @@ import { multisigs } from "../../../mocks/multisig";
 import { render } from "../../../mocks/testUtils";
 import { multisigActions } from "../../../utils/redux/slices/multisigsSlice";
 import accountsSlice from "../../../utils/redux/slices/accountsSlice";
-import { makeAccountOperations } from "../../sendForm/types";
+import { makeAccountOperations } from "../../../types/AccountOperations";
 import { contract, makeStorageJSON } from "../../../multisig/multisigContract";
 import BigNumber from "bignumber.js";
 import { DynamicModalContext } from "../../DynamicModal";
