+++ src/components/SendFlow/SendNFTRecapTile.tsx
@@ -1,9 +1,9 @@
 import { AspectRatio, Box, Flex, Heading, Image } from "@chakra-ui/react";
-import { NFT } from "../../../types/Token";
-import { thumbnailUri } from "../../../types/Token";
-import { getIPFSurl } from "../../../utils/token/nftUtils";
-import colors from "../../../style/colors";
-import { truncate } from "../../../utils/format";
+import { NFT } from "../../types/Token";
+import { thumbnailUri } from "../../types/Token";
+import { getIPFSurl } from "../../utils/token/nftUtils";
+import colors from "../../style/colors";
+import { truncate } from "../../utils/format";
 
 export const SendNFTRecapTile = ({ nft }: { nft: NFT }) => {
   const url = getIPFSurl(thumbnailUri(nft));
